Modify style.css to match the project specifications
Various declarations had to be made for each class to meet the project requirements. I learned that width is only necessary for elements when trying to horizontally center-align via the auto margin method; for sub-elements, like the third <div> center-alignment is achieved via the text-align property so there is no need to add a width.